requirements.txt
djangorestframework>=3.15
django-cors-headers>=4.3